Move `Rustc`, `Cargo` and `Rustfmt` to the `components` package
