+++ src/main/kotlin/org/rust/cargo/runconfig/CargoTestRunState.kt
@@ -19,7 +19,7 @@ import org.rust.cargo.runconfig.command.CargoCommandConfiguration
 import org.rust.cargo.runconfig.console.CargoTestConsoleBuilder
 import org.rust.cargo.toolchain.CargoCommandLine
 import org.rust.cargo.toolchain.RustChannel
-import org.rust.cargo.toolchain.RustcVersion
+import org.rust.cargo.toolchain.components.RustcVersion
 import java.time.LocalDate
 
 class CargoTestRunState(
