+++ src/main/kotlin/org/rust/cargo/runconfig/legacy/RsAsyncRunner.kt
@@ -33,9 +33,9 @@ import org.rust.cargo.runconfig.buildtool.CargoBuildManager.getBuildConfiguratio
 import org.rust.cargo.runconfig.buildtool.CargoBuildManager.isBuildConfiguration
 import org.rust.cargo.runconfig.buildtool.CargoBuildManager.isBuildToolWindowEnabled
 import org.rust.cargo.runconfig.command.CargoCommandConfiguration
-import org.rust.cargo.toolchain.Cargo
-import org.rust.cargo.toolchain.Cargo.Companion.cargoCommonPatch
 import org.rust.cargo.toolchain.CargoCommandLine
+import org.rust.cargo.toolchain.components.Cargo
+import org.rust.cargo.toolchain.components.Cargo.Companion.cargoCommonPatch
 import org.rust.cargo.toolchain.impl.CargoMetadata
 import org.rust.cargo.util.CargoArgsParser.Companion.parseArgs
 import org.rust.openapiext.saveAllDocuments
