+++ src/main/kotlin/org/rust/ide/actions/RustfmtCargoProjectAction.kt
@@ -14,8 +14,8 @@ import org.rust.cargo.project.model.CargoProject
 import org.rust.cargo.project.settings.toolchain
 import org.rust.cargo.runconfig.command.workingDirectory
 import org.rust.cargo.runconfig.getAppropriateCargoProject
-import org.rust.cargo.toolchain.Rustfmt
 import org.rust.cargo.toolchain.Rustup.Companion.checkNeedInstallRustfmt
+import org.rust.cargo.toolchain.components.Rustfmt
 import org.rust.openapiext.saveAllDocumentsAsTheyAre
 
 class RustfmtCargoProjectAction : DumbAwareAction() {
