+++ src/main/kotlin/org/rust/ide/actions/RustfmtFileAction.kt
@@ -18,8 +18,8 @@ import org.rust.cargo.project.model.CargoProject
 import org.rust.cargo.project.model.cargoProjects
 import org.rust.cargo.project.settings.toolchain
 import org.rust.cargo.runconfig.command.workingDirectory
-import org.rust.cargo.toolchain.Rustfmt
 import org.rust.cargo.toolchain.Rustup.Companion.checkNeedInstallRustfmt
+import org.rust.cargo.toolchain.components.Rustfmt
 import org.rust.lang.core.psi.isRustFile
 import org.rust.openapiext.checkWriteAccessNotAllowed
 import org.rust.openapiext.computeWithCancelableProgress
