+++ src/main/kotlin/org/rust/ide/console/RsConsoleRunner.kt
@@ -33,7 +33,7 @@ import com.intellij.util.ui.MessageCategory
 import org.rust.cargo.project.model.cargoProjects
 import org.rust.cargo.project.settings.toolchain
 import org.rust.cargo.runconfig.command.workingDirectory
-import org.rust.cargo.toolchain.Cargo
+import org.rust.cargo.toolchain.components.Cargo
 import org.rust.ide.icons.RsIcons
 import org.rust.openapiext.saveAllDocuments
 import java.awt.BorderLayout
