+++ src/main/kotlin/org/rust/ide/intentions/RunCargoExpandIntention.kt
@@ -10,8 +10,8 @@ import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
 import org.rust.cargo.project.model.CargoProject
 import org.rust.cargo.project.workspace.CargoWorkspace
-import org.rust.cargo.toolchain.Cargo.Companion.checkNeedInstallCargoExpand
 import org.rust.cargo.toolchain.CargoCommandLine
+import org.rust.cargo.toolchain.components.Cargo.Companion.checkNeedInstallCargoExpand
 import org.rust.lang.core.psi.ext.*
 import org.rust.stdext.buildList
