+++ src/main/kotlin/org/rust/ide/newProject/ui/RsNewProjectPanel.kt
@@ -24,7 +24,7 @@ import com.intellij.ui.components.Link
 import com.intellij.ui.layout.LayoutBuilder
 import com.intellij.util.ui.JBUI
 import org.rust.cargo.project.settings.ui.RustProjectSettingsPanel
-import org.rust.cargo.toolchain.Cargo
+import org.rust.cargo.toolchain.components.Cargo
 import org.rust.ide.newProject.ConfigurationData
 import org.rust.ide.newProject.RsCustomTemplate
 import org.rust.ide.newProject.RsGenericTemplate
