+++ src/test/kotlin/org/rust/cargo/runconfig/CargoTestRunStateTest.kt
@@ -13,7 +13,7 @@ import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import org.rust.cargo.toolchain.CargoCommandLine
 import org.rust.cargo.toolchain.RustChannel
-import org.rust.cargo.toolchain.RustcVersion
+import org.rust.cargo.toolchain.components.RustcVersion
 import java.nio.file.Path
 import java.nio.file.Paths
 import java.time.LocalDate
