+++ src/test/kotlin/org/rust/cargo/toolchain/RustcVersionParsingTest.kt
@@ -10,6 +10,8 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import org.rust.cargo.toolchain.RustChannel.*
+import org.rust.cargo.toolchain.components.RustcVersion
+import org.rust.cargo.toolchain.components.parseRustcVersion
 import kotlin.test.assertEquals
 import java.time.LocalDate.parse as parseDate
